Kafka-streaming/requirements.txt
ansi2html==1.8.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
pandas==2.0.3
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vaderSentiment==3.3.2
werkzeug==3.0.1
zipp==3.17.0
